TimeTable: change departure board from_datetime type
datetime must be handled as posix time stamp from jormungandr

change a narrowing conversion from uint32 to uint64 in utils_test

+++ source/routing/tests/routing_api_test.cpp
@@ -611,8 +611,8 @@ BOOST_AUTO_TEST_CASE(journey_array){
 
     navitia::georef::StreetNetwork sn_worker(*data.geo_ref);
 
-    // On met les horaires dans le desordre pour voir s'ils sont bien triÃ© comme attendu
-    std::vector<uint32_t> datetimes({ntest::to_posix_timestamp("20120614T080000"), ntest::to_posix_timestamp("20120614T090000")});
+    //we put the time not in the right order to check that they are correctly sorted
+    std::vector<uint64_t> datetimes({ntest::to_posix_timestamp("20120614T080000"), ntest::to_posix_timestamp("20120614T090000")});
     pbnavitia::Response resp = nr::make_response(raptor, origin, destination, datetimes, true,
                                                  navitia::type::AccessibiliteParams()/*false*/, forbidden, sn_worker, false, true);
